backend/requirements.txt
# Core web stack
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1

# ASGI / websocket
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1

# Database
mysqlclient==2.2.7

# Common app libs
requests==2.32.5
beautifulsoup4==4.14.3
pillow==12.1.0

# RAG / vector db
chromadb==1.1.1
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-text-splitters==0.3.11
pypdf==6.1.3
docx2txt==0.9
sentence-transformers==3.0.1
